+++ go/tryableonce/once_test.go
@@ -1,4 +1,4 @@
-package main
+package tryableonce
 
 import (
 	"sync"
